[fix]Java does not have an unsigned integer type, so it will interpret negative numbers.
Converter.java    Fixed sign extension issues in byteArrayToInt and byteArrayToDouble, added toByteArray(Long) and byteArrayToUnsignedInt methods
IntResponse.java    Changed value type from int to long
SocketService.java    sendAndReceiveInt() now returns Long
BankingService.java    openAccount() now returns Long
FieldDefn.java    Changed ACCOUNT_NO and DEST_ACCOUNT_NO types to Long.class
MySerializable.java    Added serialization support for Long type
UpdateBalanceRequest.java    accountNumber changed to Long, fixed existing bug in getFieldDefs
TransferRequest.java    accountNumber/destAccountNumber changed to Long
CloseAccountRequest.java    accountNumber changed to Long
GetBalanceRequest.java    accountNumber changed to Long
BankingCli.java    Use Long.parseLong() to parse account numbers

Translated with DeepL.com (free version)

+++ client/src/main/java/com/chencraft/ntu/cli/BankingCli.java
@@ -96,7 +96,7 @@ public class BankingCli {
         request.setCurrency(Currency.valueOf(parts[3].toUpperCase()));
         request.setInitialBalance(Double.parseDouble(parts[4]));
 
-        Integer accountNumber = bankingService.openAccount(request);
+        Long accountNumber = bankingService.openAccount(request);
         System.out.println("[SUCCESS] Account opened successfully. Account Number: " + accountNumber);
     }
 
@@ -107,7 +107,7 @@ public class BankingCli {
         }
         CloseAccountRequest request = new CloseAccountRequest();
         request.setName(parts[1]);
-        request.setAccountNumber(Integer.parseInt(parts[2]));
+        request.setAccountNumber(Long.parseLong(parts[2]));
         request.setPassword(parts[3]);
 
         String msg = bankingService.closeAccount(request);
@@ -121,7 +121,7 @@ public class BankingCli {
         }
         UpdateBalanceRequest request = new UpdateBalanceRequest();
         request.setName(parts[1]);
-        request.setAccountNumber(Integer.parseInt(parts[2]));
+        request.setAccountNumber(Long.parseLong(parts[2]));
         request.setPassword(parts[3]);
         request.setCurrency(Currency.valueOf(parts[4].toUpperCase()));
         request.setAmount(Double.parseDouble(parts[5]));
@@ -138,7 +138,7 @@ public class BankingCli {
         }
         UpdateBalanceRequest request = new UpdateBalanceRequest();
         request.setName(parts[1]);
-        request.setAccountNumber(Integer.parseInt(parts[2]));
+        request.setAccountNumber(Long.parseLong(parts[2]));
         request.setPassword(parts[3]);
         request.setCurrency(Currency.valueOf(parts[4].toUpperCase()));
         request.setAmount(Double.parseDouble(parts[5]));
@@ -155,7 +155,7 @@ public class BankingCli {
         }
         GetBalanceRequest request = new GetBalanceRequest();
         request.setName(parts[1]);
-        request.setAccountNumber(Integer.parseInt(parts[2]));
+        request.setAccountNumber(Long.parseLong(parts[2]));
         request.setPassword(parts[3]);
 
         Double balance = bankingService.getBalance(request);
@@ -169,9 +169,9 @@ public class BankingCli {
         }
         TransferRequest request = new TransferRequest();
         request.setName(parts[1]);
-        request.setAccountNumber(Integer.parseInt(parts[2]));
+        request.setAccountNumber(Long.parseLong(parts[2]));
         request.setPassword(parts[3]);
-        request.setDestAccountNumber(Integer.parseInt(parts[4]));
+        request.setDestAccountNumber(Long.parseLong(parts[4]));
         request.setCurrency(Currency.valueOf(parts[5].toUpperCase()));
         request.setAmount(Double.parseDouble(parts[6]));
